test: Moved real tests to testreal and query text JUnit tests to testtext
