+++ library/src/test/java/com/alexfu/sqlitequerybuilder/testtext/DropTest.java
@@ -1,4 +1,4 @@
-package com.alexfu.sqlitequerybuilder;
+package com.alexfu.sqlitequerybuilder.testtext;
 
 import org.junit.Test;
